apps/clip-service/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
torch>=2.0.0
transformers>=4.35.0
pillow>=10.0.0
requests>=2.31.0
numpy>=1.24.0
python-multipart>=0.0.6
